pyproject.toml
[project]
name = "sums"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "ebooklib>=0.19",
    "google-genai>=1.13.0",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
